requirements.txt
discord.py-self>=2.0.1
typing-extensions>=4.0.0
PySide6==6.8.0.2
pyinstaller==6.3.0
